Pie Infocomm internship project
project is basically use to recognise the emotion in speech. I have used 4 types of data sets with are available on https://www.kaggle.com/
Types of dataset:
1. CREMA
2. RAVDESS
3. SAVEE
4. TESS
The above mentioned datasets consists mixture of emotion audios which are helpfull for this project.
There is also a demonstration video to help out in coding section.